Rename `ActionsModal` to `ActionMenu`
