+++ ui/src/groups/NewGroup/NewGroupInvite.tsx
@@ -12,7 +12,7 @@ import { PrivacyType } from '@/types/groups';
 import { Status } from '@/logic/status';
 import LoadingSpinner from '@/components/LoadingSpinner/LoadingSpinner';
 import X16Icon from '@/components/icons/X16Icon';
-import ActionsModal, { Action } from '@/components/ActionsModal';
+import ActionMenu, { Action } from '@/components/ActionMenu';
 
 interface NewGroupInviteProps {
   groupName: string;
@@ -52,7 +52,7 @@ function MemberRoleDropDownMenu({
   }));
 
   return (
-    <ActionsModal open={isOpen} onOpenChange={setIsOpen} actions={actions}>
+    <ActionMenu open={isOpen} onOpenChange={setIsOpen} actions={actions}>
       <button
         className={cn(
           'default-focus text-md mx-2 flex items-center rounded-lg bg-gray-400 py-0.5 px-2 font-bold text-black mix-blend-multiply transition-opacity focus-within:opacity-100 hover:opacity-100 group-focus-within:opacity-100 group-hover:opacity-100 dark:mix-blend-screen'
@@ -62,7 +62,7 @@ function MemberRoleDropDownMenu({
         <span>{selectedRole}</span>
         <CaretDownIcon className="h-4 w-4" />
       </button>
-    </ActionsModal>
+    </ActionMenu>
   );
 }
